--- backlog/ModelIndex/ModelIndexTable/constants.ts
@@ -1,1 +1,0 @@
-export const ACTION_COLUMN = '__ACTION_COLUMN__';
