--- backlog/ModelIndex/ModelIndexTable/index.ts
@@ -1,14 +1,0 @@
-// components
-export * from './ModelIndexTable';
-export * from './ModelIndexTableActionCell';
-export * from './ModelIndexTableActionHeaderCell';
-export * from './ModelIndexTableBody';
-export * from './ModelIndexTableCell';
-export * from './ModelIndexTableErrorRow';
-export * from './ModelIndexTableHead';
-export * from './ModelIndexTableHeaderCell';
-export * from './ModelIndexTableHeaderRow';
-export * from './ModelIndexTableRow';
-
-// constants
-export * from './constants';
