--- backlog/ModelIndex/index.ts
@@ -1,16 +1,0 @@
-// components
-export * from './ModelIndex';
-export * from './ModelIndexPagination';
-export * from './ModelIndexSettings';
-export * from './ModelIndexTable';
-export * from './ModelIndexTitle';
-export * from './SortWrapper';
-
-// contexts
-export * from './ModelIndexStoreContext';
-
-// hooks
-export * from './useModelIndexStore';
-
-// utils
-export * from './utils';
