--- backlog/ModelIndexSettings/index.ts
@@ -1,2 +1,0 @@
-export * from './ModelIndexSettings';
-// export * from './ModelIndexSortSetting';
