+++ src/ModelTable/ModelTableHeaderRow.tsx
@@ -3,6 +3,8 @@ import { DnDSortableContextWrapper } from '@/utils';
 
 import { ModelTableHead } from './ModelTableHead';
 import { useModelTableStore } from './useModelTableStore';
+import { ACTION_COLUMN } from './ModelTable';
+import { ModelTableActionHead } from './ModelTableActionHead';
 
 export interface ModelTableHeaderRowProps extends TableHeaderRowProps {}
 
@@ -18,18 +20,20 @@ export const ModelTableHeaderRow = ({
     (state) => state.tableOptions?.draggable ?? true,
   );
   return (
-    <DnDSortableContextWrapper draggable={draggable} dndList={renderedFields}>
-      <TableHeaderRow prefilled={prefilled} {...tableHeaderRowProps}>
+    <TableHeaderRow prefilled={prefilled} {...tableHeaderRowProps}>
+      <DnDSortableContextWrapper draggable={draggable} dndList={renderedFields}>
         {children === undefined ? (
           <>
-            {renderedFields.map((field) => (
-              <ModelTableHead key={field} field={field} />
-            ))}
+            {renderedFields.map((field) => {
+              if (field === ACTION_COLUMN)
+                return <ModelTableActionHead key={ACTION_COLUMN} />;
+              return <ModelTableHead key={field} field={field} />;
+            })}
           </>
         ) : (
           children
         )}
-      </TableHeaderRow>
-    </DnDSortableContextWrapper>
+      </DnDSortableContextWrapper>
+    </TableHeaderRow>
   );
 };
